feat(gulp): add watch task
For now, when any file changes, run test task (which runs all tasks).

There's a bug somewhere in the test task process that is changing the
current working directory and not changing it back. In the end event, the
CWD is changed back to the original.

In the future, only changed files should be evaluated instead of everything.