chore(deps): update Babel
Convert app/index to export an object instead of a class.

Calling super(...args) was throwing compile errors saying super.*
was not allowed to be used before super(). Not sure if this is a
Babel bug or fix. Either, I've been wanting to convert back to
exporting an object, so it was done to fix the upgrade process.